MiniBar/requirements.txt
fastapi[standard]
uvicorn
psycopg2-binary
SQLAlchemy